Add celery task to calculate integration statuses
