+++ cdip_admin/integrations/models/v2/__init__.py
@@ -15,7 +15,7 @@ from .models import (
     SourceState,
     GundiTrace,
     IntegrationWebhook,
-    WebhookConfiguration,
+    WebhookConfiguration
 )
 from .services import *
 from .source_filters import *
